requirements.txt
langchain
faiss-cpu
google-generativeai
pypdf
pandas
python-dotenv
streamlit

# For Jupyter notebook support
ipykernel
